HUD implemented. includes player healthbar.
Player taking damage implemented. Includes an invincibility frames feature, to avoid damage being too frequent/instakilling the player before they can be knocked back out of the enemy attack.

Player attacking partially implemented. Needs hitbox added, and sprites updated so attacking looks better.

Array made for which tiles are collidable. A short loop checks through the array to check if a tile is water, or a wall, for example, and therefore if it is a collidable tile.
This allows us to contiue using tileMap only, not needing to make one for "tileRules" along with each map/level.